# api-docs.yaml
openapi: 3.1.0
info:
  title: GBCC Website API
  description: |
    REST API интернет-магазина / каталога GBCC: аутентификация (JWT в HttpOnly cookies),
    аккаунты, каталог товаров и файлы, новости, акции (скидки), заказы, обращения в поддержку.

    **Авторизация в Swagger UI:** схема `accessTokenCookie` — cookie `GBCC_ACCESS_TOKEN`
    (после успешного `POST /auth/login` в том же браузере cookie подставляется автоматически;
    при тестах из Postman — передайте cookie вручную).

    **Гостевой токен поддержки:** схема `supportGuestToken` — заголовок `X-Support-Token`
    (значение из `guestAccessToken` после создания обращения без входа).
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Local Dev Server
...
paths:
...
  /support/conversations:
    get:
      tags:
      - Support
      summary: Список обращений (администраторы)
      description: "Требует JWT. В сервисе допускаются только ADMIN и OWNER. Фильт\
        ры: status, query (тема или email гостя)."
      operationId: searchForAdmin
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - CLOSED
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Страница
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      tags:
      - Support
      summary: Создать обращение
      description: |
        Без JWT: обязательны message, guestName, guestEmail (валидный email).
        С JWT и ролью CUSTOMER: только message (и опционально subject); гостевой токен не выдаётся.
        Администраторам создавать диалог через этот метод нельзя — отвечают в существующем.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportConversationRequestDto"
        required: true
      responses:
        "201":
          description: Создано; для гостя в теле guestAccessToken
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateSupportConversationResponseDto"
        "400":
          description: Некорректные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Попытка создать от имени администратора
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /cooperation-requests:
    get:
      tags:
      - Cooperation requests
      summary: Список заявок (администраторы)
      description: "JWT. Фильтры: onlyUnassigned, query (поиск по имени, email, те\
        лефону, комментарию, типу сотрудничества)."
      operationId: searchForAdmin_1
      parameters:
      - name: onlyUnassigned
        in: query
        required: false
        schema:
          type: boolean
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Страница заявок
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      tags:
      - Cooperation requests
      summary: Отправить заявку
      description: Публично. Требуется согласие на обработку персональных данных.
      operationId: create_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCooperationRequestDto"
        required: true
      responses:
        "201":
          description: Создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CooperationRequestResponseDto"
        "400":
          description: Валидация или файл не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /cooperation-requests/{id}/release:
    post:
      tags:
      - Cooperation requests
      summary: Снять назначение
      description: JWT. Снимает привязку администратора; может назначенный админ или
        OWNER.
      operationId: release
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Снято
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CooperationRequestResponseDto"
        "400":
          description: Не было назначения
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет прав
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
  /contact-requests:
    get:
      tags:
      - Contact requests
      summary: Список заявок (администраторы)
      description: "JWT. Фильтры: onlyUnassigned (только без назначенного), query\
        \ (поиск по имени, email, телефону, тексту сообщения)."
      operationId: searchForAdmin_2
      parameters:
      - name: onlyUnassigned
        in: query
        required: false
        schema:
          type: boolean
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Страница заявок
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      tags:
      - Contact requests
      summary: Отправить заявку
      description: Публично. Требуется согласие на обработку персональных данных.
      operationId: create_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequestDto"
        required: true
      responses:
        "201":
          description: Создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactRequestResponseDto"
        "400":
          description: Валидация
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /contact-requests/{id}/take:
    post:
      tags:
      - Contact requests
      summary: Взять заявку на изучение
      description: JWT. Назначает текущего администратора. Если уже назначен другой
        — 409 (OWNER может переназначить на себя).
      operationId: take_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Назначено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactRequestResponseDto"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Уже назначена другому
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
  /contact-requests/{id}/release:
    post:
      tags:
      - Contact requests
      summary: Снять назначение
      description: JWT. Снимает привязку администратора; может назначенный админ или
        OWNER.
      operationId: release_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Снято
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactRequestResponseDto"
        "400":
          description: Не было назначения
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет прав
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
...
  /contact-requests/{id}:
    get:
      tags:
      - Contact requests
      summary: Заявка по id (администраторы)
      description: JWT. Полная карточка с назначением.
      operationId: getById_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactRequestResponseDto"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
...
components:
  schemas:
...
    ApiErrorField:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки API (см. GlobalExceptionHandler)
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorField"
...
    CreateSupportConversationRequestDto:
      type: object
      properties:
        subject:
          type: string
          maxLength: 500
          minLength: 0
        message:
          type: string
          maxLength: 16000
          minLength: 0
        guestName:
          type: string
          maxLength: 255
          minLength: 0
        guestEmail:
          type: string
          format: email
          maxLength: 255
          minLength: 0
      required:
      - message
    CreateSupportConversationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        guestAccessToken:
          type: string
        subject:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
        createdAt:
          type: string
          format: date-time
...
    CreateCooperationRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        phone:
          type: string
          maxLength: 64
          minLength: 0
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        cooperationType:
          type: string
          maxLength: 128
          minLength: 0
        comment:
          type: string
        attachmentFileId:
          type: string
          format: uuid
        consentProcessing:
          type: boolean
      required:
      - email
      - name
      - phone
    CooperationRequestResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        cooperationType:
          type: string
        comment:
          type: string
        attachmentFileId:
          type: string
          format: uuid
        consentProcessing:
          type: boolean
        assignedToAccountId:
          type: string
          format: uuid
        assignedToName:
          type: string
        assignedToEmail:
          type: string
        assignedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateContactRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        phone:
          type: string
          maxLength: 64
          minLength: 0
        message:
          type: string
          minLength: 1
        consentProcessing:
          type: boolean
      required:
      - email
      - message
      - name
    ContactRequestResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        consentProcessing:
          type: boolean
        assignedToAccountId:
          type: string
          format: uuid
        assignedToName:
          type: string
        assignedToEmail:
          type: string
        assignedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      description: |
        JWT access token (HttpOnly cookie). Выставляется после POST /auth/login.
        В Swagger UI: Authorize → значение токена (если копируете из DevTools) или вход через браузер на том же origin.
      name: GBCC_ACCESS_TOKEN
      in: cookie
    supportGuestToken:
      type: apiKey
      description: |
        Токен гостя для обращения в поддержку без аккаунта.
        Возвращается в теле ответа POST /support/conversations (поле guestAccessToken).
      name: X-Support-Token
      in: header